refactor(i18n): isolate locale utils and normalize server cookie decode
