+++ frontend/src/core/i18n/server.ts
@@ -1,9 +1,17 @@
 import { cookies } from "next/headers";
 
-import { normalizeLocale, type Locale } from "./index";
+import { normalizeLocale, type Locale } from "./locale";
 
 export async function detectLocaleServer(): Promise<Locale> {
   const cookieStore = await cookies();
-  const locale = cookieStore.get("locale")?.value;
+  let locale = cookieStore.get("locale")?.value;
+  if (locale !== undefined) {
+    try {
+      locale = decodeURIComponent(locale);
+    } catch {
+      // Keep raw cookie value when decoding fails.
+    }
+  }
+
   return normalizeLocale(locale);
 }
